requirements.txt
tensorflow
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
google-api-core
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
numpy
pandas
scikit-learn
fastapi
uvicorn